api/requirements.txt
numpy<2
faiss-cpu==1.8.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
psycopg[binary]==3.2.1
pydantic==2.8.2
requests==2.32.3
python-dateutil>=2.8
word2number
packaging
